+++ app/src/main/java/com/joshuarichardson/fivewaystowellbeing/storage/dao/ActivityRecordDao.java
@@ -17,10 +17,10 @@ public interface ActivityRecordDao {
     long insert(ActivityRecord activityRecord);
 
     @Query("SELECT * FROM activity_records WHERE id = :activityRecordId")
-    LiveData<List<ActivityRecord>> getActivityRecordById(int activityRecordId);
+    LiveData<List<ActivityRecord>> getActivityRecordById(long activityRecordId);
 
     @Query("SELECT * FROM activity_records WHERE timestamp >= :startTime AND timestamp <= :endTime")
-    LiveData<List<ActivityRecord>> getActivitiesInTimeRange(int startTime, int endTime);
+    LiveData<List<ActivityRecord>> getActivitiesInTimeRange(long startTime, long endTime);
 
     @Query("SELECT * FROM activity_records")
     LiveData<List<ActivityRecord>> getAllActivities();
