+++ app/src/main/java/com/joshuarichardson/fivewaystowellbeing/storage/dao/SurveyResponseDao.java
@@ -17,13 +17,13 @@ public interface SurveyResponseDao {
     long insert(SurveyResponse surveyResponse);
 
     @Query("SELECT * FROM survey_response WHERE id = :surveyId")
-    LiveData<List<SurveyResponse>> getSurveyResponseById(int surveyId);
+    LiveData<List<SurveyResponse>> getSurveyResponseById(long surveyId);
 
     @Query("SELECT * FROM survey_response")
     LiveData<List<SurveyResponse>> getAllSurveyResponses();
 
     @Query("SELECT * FROM survey_response WHERE timestamp >= :startTimestamp AND timestamp <= :endTimestamp")
-    LiveData<List<SurveyResponse>> getSurveyResponsesByTimestampRange(int startTimestamp, int endTimestamp);
+    LiveData<List<SurveyResponse>> getSurveyResponsesByTimestampRange(long startTimestamp, long endTimestamp);
 
     // ToDo Will need to add a delete
 }
